# backend/target/classes/static/apiDefinition.yaml
openapi: 3.0.0
info:
  title: Instagram Follower Count API
  description: API to extract follower count from Instagram profiles using Java and JSON files format.
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Development server
tags:
    - name: file
      description: JSON data uploaded via files.
    - name: stringFormat
      description: JSON data sent via string message.
paths:
  /uploadStringFormat:
    post:
      tags:
        - stringFormat
      summary: Upload followers and following data in string JSON format
      description: This method accepts two JSON strings, one for followers and one for following, and processes them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonStrings'
      responses:
        '200':
          description: Successful upload
        '400':
          description: An empty string, one with syntax error/s or wrong form-data was sent.
        '405':
          description: A non-POST request was sent, while this method supports only POST requests.
  /uploadFollowers:
    post:
      tags:
        - file
      summary: Upload following
      description: This method accepts a JSON file derived from Instagram containing the list of people who you follow and processes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Followers'
      responses:
        '200':
          description: Successful upload
        '400':
          description: An empty file, one with syntax error/s or wrong form-data was sent.
        '405':
          description: A non-POST request was sent, while this method supports only POST requests.
        '413':
          description: The maximum size of a file to upload is 5 MB, uploading a larger file leads to this response code.
        '415':
          description: You have sent a POST request different from the form-data media type.
  /uploadFollowing:
    post:
      tags:
        - file
      summary: Upload following
      description: This method accepts a JSON file derived from Instagram containing the list of people who you follow and processes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Following'
      responses:
        '200':
          description: Successful upload
        '400':
          description: An empty file, one with syntax error/s or wrong form-data was sent.
        '405':
          description: A non-POST request was sent, while this method supports only POST requests.
        '413':
          description: The maximum size of a file to upload is 5 MB, uploading a larger file leads to this response code.
        '415':
          description: You have sent a POST request different from the form-data media type.
components:
  schemas:
    JsonStrings:
      type: object
      properties:
        jsonFollowers:
          type: string
          description: JSON string of followers data
          example: "[{ \"title\": \"\", \"media_list_data\": [], \"string_list_data\": [{ \"href\": \"https://www.instagram.com/someinstagram\", \"value\": \"someinstagram\", \"timestamp\": 1714580639 }]}]"
        jsonFollowing:
          type: string
          description: JSON string of following data
          example: "{ \"relationships_following\": [{ \"title\": \"\", \"media_list_data\": [], \"string_list_data\": [{ \"href\": \"https://www.instagram.com/someinstagram\", \"value\": \"someinstagram\", \"timestamp\": 1714580639 }] }] }"
    InstagramData:
      type: object
      properties:
        title:
          type: string
        media_list_data:
          type: array
          items: { }
        string_list_data:
          type: array
          items:
            type: object
            properties:
              href:
                type: string
                value: "https://www.instagram.com/someinstagram"
              value:
                type: string
                value: "someinstagram"
              timestamp:
                type: integer
                value: 1716396721
    Followers:
      type: array
      items:
        $ref: '#/components/schemas/InstagramData'
      example:
        - title: ""
          media_list_data: [ ]
          string_list_data:
            - href: "https://www.instagram.com/someinstagram"
              value: "someinstagram"
              timestamp: 1716396721
    Following:
      type: object
      properties:
        relationships_following:
          type: array
          items:
            $ref: '#/components/schemas/InstagramData'
      example:
        relationships_following:
          - title: ""
            media_list_data: [ ]
            string_list_data:
              - href: "https://www.instagram.com/someinstagram"
                value: "someinstagram"
                timestamp: 1716396721
securitySchemes:
  BasicAuth:
    type: http
    scheme: basic
security:
  - BasicAuth: [ ]